v0.5.6：悬浮条中/En 全局语义 + 语音入口 + 热键门控接线 + 全工作区 clippy/fmt 清零
在途 v0.5.6 收尾为可发布最终形态：

- 算法服务 ascii_mode 改全局语义：任一会话切换后所有应用下一键跟上，
  悬浮条显示/切换全局中英态（GetOption/SetOption/ToggleAscii 走 GLOBAL_ASCII）
- 悬浮条麦克风经 WM_APP_SPEECH_TOGGLE 真触发语音转写面板（与 Ctrl+Shift+S 同入口）；
  移除重复网格入口；中/En 与拼/双切换改乐观更新（失败回滚）
- 附录 K 热键门控接线兑现：ai_panel::register_hotkey 读取
  shurufa_options::hotkey_gates()，listener SetTimer 2s 轮询变化时反注册
  重注册（原子位图去重；默认全开行为不变）
- 全工作区 cargo clippy --all-targets -- -D warnings 清零（约 60 项：
  死代码 MRU_ENABLED/onscreen_kbd::vk 移除、speech Failed/Flush 与 skin
  滚动条段按工程理由豁免、QPC unsafe 保留、GPU 枚举 large_enum_variant
  豁免附注释、文档列表/空行/冗余转换等机械项）
- cargo fmt --all 全量规范化（此前 fmt --check 已有历史漂移）
- 设置中心新增 ime_mode_status/ime_mode_toggle/trigger_speech 命令；
  版本统一 0.5.6/17；README/CHANGELOG/迭代方向/验证报告同步更新；
  .gitignore 排除 DSH 工具工作区

+++ core/ime-ipc/src/pipe.rs
@@ -9,7 +9,7 @@ use std::io;
 use std::ffi::c_void;
 
 use windows::core::HSTRING;
-use windows::Win32::Foundation::{CloseHandle, GENERIC_READ, GENERIC_WRITE, HANDLE, LocalFree};
+use windows::Win32::Foundation::{CloseHandle, LocalFree, GENERIC_READ, GENERIC_WRITE, HANDLE};
 use windows::Win32::Security::Authorization::{
     ConvertStringSecurityDescriptorToSecurityDescriptorW, SDDL_REVISION_1,
 };
@@ -19,9 +19,9 @@ use windows::Win32::Storage::FileSystem::{
     PIPE_ACCESS_DUPLEX,
 };
 use windows::Win32::System::Pipes::{
-    ConnectNamedPipe, CreateNamedPipeW, DisconnectNamedPipe, PeekNamedPipe, SetNamedPipeHandleState,
-    PIPE_READMODE_MESSAGE, PIPE_TYPE_MESSAGE,
-    PIPE_UNLIMITED_INSTANCES, PIPE_WAIT,
+    ConnectNamedPipe, CreateNamedPipeW, DisconnectNamedPipe, PeekNamedPipe,
+    SetNamedPipeHandleState, PIPE_READMODE_MESSAGE, PIPE_TYPE_MESSAGE, PIPE_UNLIMITED_INSTANCES,
+    PIPE_WAIT,
 };
 
 pub const PIPE_NAME: &str = r"\\.\pipe\shurufa-algo";
@@ -32,7 +32,7 @@ const MAX_FRAME: u32 = crate::MAX_FRAME_BYTES as u32;
 type IoResult<T> = Result<T, io::Error>;
 
 fn win_err(e: windows::core::Error) -> io::Error {
-    io::Error::from_raw_os_error(e.code().0 as i32)
+    io::Error::from_raw_os_error(e.code().0)
 }
 
 /// 服务端：新建一个命名管道实例并进入监听态。
@@ -48,8 +48,7 @@ unsafe impl Send for PipeServer {}
 /// （如安装器以管理员启动服务），High 进程创建的管道默认隐式 High 标签会拒绝
 /// 普通（Medium）应用连接 → 输入法整体失效（2026-08-14 实机复现：安装后
 /// 普通应用 IPC 全拒，err=5）。显式 Medium 标签让普通应用在 DACL 允许时始终可达。
-const PIPE_SDDL: &str =
-    "D:(A;;GA;;;SY)(A;;GA;;;BA)(A;;GA;;;OW)S:(ML;;NW;;;ME)";
+const PIPE_SDDL: &str = "D:(A;;GA;;;SY)(A;;GA;;;BA)(A;;GA;;;OW)S:(ML;;NW;;;ME)";
 
 impl PipeServer {
     /// 新建管道实例并允许客户端连接（FILE_FLAG_FIRST_PIPE_INSTANCE 仅首个实例）。
@@ -100,9 +99,7 @@ impl PipeServer {
                 Ok(()) => Ok(()),
                 Err(e) => {
                     // 客户端在服务端 ConnectNamedPipe 之前已连接：视为成功
-                    if e.code().0 as i32
-                        == windows::Win32::Foundation::ERROR_PIPE_CONNECTED.0 as i32
-                    {
+                    if e.code().0 == windows::Win32::Foundation::ERROR_PIPE_CONNECTED.0 as i32 {
                         return Ok(());
                     }
                     Err(win_err(e))
